+++ Block/Adminhtml/Form/Edit/Tab/Fields/Form.php
@@ -108,6 +108,12 @@ class Form extends \Alekseon\AlekseonEav\Block\Adminhtml\Entity\Edit\Form
      */
     protected function addFieldFieldset($form, $formFieldId, $settings = [])
     {
+        if (isset($settings['attribute'])) {
+            $attribute = $settings['attribute'];
+        } else {
+            $attribute = false;
+        }
+
         if (isset($settings['is_new_field']) && $settings['is_new_field']) {
             $isNewField = true;
         } else {
@@ -168,15 +174,17 @@ class Form extends \Alekseon\AlekseonEav\Block\Adminhtml\Entity\Edit\Form
             )->addCustomAttribute("data-fieldcode", "frontend_input");
         }
 
-        $fieldset->addField('form_field_' . $formFieldId . '_is_required', 'select',
-            [
-                'label' => __('Is Required'),
-                'name' => 'form_fields[' . $formFieldId . '][is_required]',
-                'values' => $this->yesNoSource->toOptionArray()
-            ]
-        )->addCustomAttribute("data-fieldcode", "is_required");
+        if (!$attribute || $attribute->getIsRequiredEditable()) {
+            $fieldset->addField('form_field_' . $formFieldId . '_is_required', 'select',
+                [
+                    'label' => __('Is Required'),
+                    'name' => 'form_fields[' . $formFieldId . '][is_required]',
+                    'values' => $this->yesNoSource->toOptionArray()
+                ]
+            )->addCustomAttribute("data-fieldcode", "is_required");
+        }
 
-        if ($attribute && $this->canSelectOptionSource($attribute)) {
+        if ($this->canSelectOptionSource($attribute)) {
             $fieldset->addField('form_field_' . $formFieldId . '_option_source_code', 'select',
                 [
                     'label' => __('Options Source'),