+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 ### Added
 
+## [Unreleased]
+### Added
+- possibility to disable form field
+- check if "is required" field is editable
+
 ## [102.2.2] - 2022-11-19
 ### Fixed
 - fix setUp() declaration in tests